# reports/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for report management",
        "title": "Reports API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/reports/employees-report": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reports"
                ],
                "summary": "Get employees report",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.ReportResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.ReportResponse": {
            "type": "object",
            "properties": {
                "employees": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_handler_dto.EmployeeResponse"
                    }
                },
                "is_fired": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "internal_handler_dto.EmployeeResponse": {
            "type": "object",
            "properties": {
                "birth_date": {
                    "type": "string"
                },
                "fio": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_fired": {
                    "type": "boolean"
                }
            }
        },
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}